IA_Demo_Test_Accidente_Carros/requirements.txt
fastapi
uvicorn
streamlit
requests
Pillow
